pyproject.toml
[tool.poetry]
name = "django-rest-api"
version = "0.1.0"
description = ""
authors = ["pierre kasongo <pierrekasongo88@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.2.3"
djangorestframework = "^3.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
